refactoring. move all gql queries to @/graphql/client. move all form components to @/components/forms. add some styling to dashboard component. make user context provider it's own component and move to _app
