+++ graphql/client/dashboard/index.ts
@@ -1,0 +1,3 @@
+export * from "./registerMutation"
+export * from "./dashboardQuery"
+export * from "./wordcloudQuery"
