+++ graphql/client/dashboard/wordcloudQuery.ts
@@ -1,0 +1,14 @@
+import { graphql } from "@/generated"
+
+export const WORDCLOUD_QUERY = graphql(`
+  query WordcloudQuery($user: Int!) {
+    dashboard(user: $user) {
+      dashboard {
+        thirtyDayWordcloud
+        sixtyDayWordcloud
+        ninetyDayWordcloud
+        yearWordcloud
+      }
+    }
+  }
+`)
