+++ graphql/client/index.ts
@@ -1,0 +1,3 @@
+export * from "./dashboard"
+export * from "./track"
+export * from "./global"
