+++ pages/[id].tsx
@@ -3,24 +3,13 @@ import Head from "next/head"
 import { useSession } from "next-auth/react"
 import { Box } from "@mui/material"
 import { useQuery } from "@apollo/client"
-import { graphql } from "@/generated"
 import Wordcloud from "@/components/wordcloud"
+import { WORDCLOUD_QUERY } from "@/graphql/client"
 
 export default function WordcloudPage() {
   const session = useSession()
   const { data: dashboardData, loading: dashboardLoading } = useQuery(
-    graphql(`
-      query WordcloudQuery($user: Int!) {
-        dashboard(user: $user) {
-          dashboard {
-            thirtyDayWordcloud
-            sixtyDayWordcloud
-            ninetyDayWordcloud
-            yearWordcloud
-          }
-        }
-      }
-    `),
+    WORDCLOUD_QUERY,
     { variables: { user: Number(session.data?.user.id) } }
   )
