+++ pages/_app.tsx
@@ -1,10 +1,11 @@
 import type { AppProps } from "next/app"
 import { ApolloProvider } from "@apollo/client"
 import Layout from "@/components/layout"
-import client from "../graphql/apolloClient"
+import client from "../graphql/client/apolloClient"
 import { ThemeProvider } from "@/components/themeProvider"
 import { SessionProvider } from "next-auth/react"
 import "animate.css"
+import { UserContextProvider } from "@/context/providers"
 
 export default function App({
   Component,
@@ -12,13 +13,15 @@ export default function App({
 }: AppProps) {
   return (
     <SessionProvider session={session}>
-      <ThemeProvider>
-        <ApolloProvider client={client}>
-          <Layout>
-            <Component {...pageProps} />
-          </Layout>
-        </ApolloProvider>
-      </ThemeProvider>
+      <ApolloProvider client={client}>
+        <UserContextProvider>
+          <ThemeProvider>
+            <Layout>
+              <Component {...pageProps} />
+            </Layout>
+          </ThemeProvider>
+        </UserContextProvider>
+      </ApolloProvider>
     </SessionProvider>
   )
 }
